Moved leptons and jets function outside of objects lib
