+++ config/parton_matching/config.py
@@ -9,7 +9,7 @@ cfg =  {
         "jsons": ["datasets/signal_ttHTobb_2018_local.json",
                   "datasets/backgrounds_MC_2018_local.json"],
         "filter" : {
-            "samples": ["ttHTobb","TTTo2L2Nu"],
+            "samples": ["TTTo2L2Nu"],
             "samples_exclude" : [],
             "year": ["2018"]
         }
@@ -21,14 +21,14 @@ cfg =  {
 
     # Executor parameters
     "run_options" : {
-        "executor"       : "futures",
-        "workers"        : 10,
-        "scaleout"       : 10,
+        "executor"       : "iterative",
+        "workers"        : 5,
+        "scaleout"       : 5,
         "partition"      : "standard",
         "walltime"       : "12:00:00",
         "mem_per_worker" : None, # GB
         "exclusive"      : True,
-        "chunk"          : 100000,
+        "chunk"          : 250000,
         "max"            : None,
         "skipbadfiles"   : None,
         "voms"           : None,
